pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arxiv-llm"
version = "0.1.0"
description = "An AI-powered semantic search tool for arXiv papers"
readme = "README.md"
authors = [{ name = "Julian Chan", email = "julianchan8897@gmail.com" }]
license = { text = "MIT" }
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "streamlit>=1.28.0",
    "arxiv>=1.4.0",
    "pandas>=1.5.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.21.0",
    "faiss-cpu>=1.7.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/arxiv-llm"
Repository = "https://github.com/yourusername/arxiv-llm.git"
Issues = "https://github.com/yourusername/arxiv-llm/issues"

[project.scripts]
arxiv-search = "scripts.run_arxiv_search:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md"]

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = ["F401"]
